// src/result/types.ts
export type ResultProps = {
    /**
     * Статус результата, который определяет визуальное представление.
     * Может принимать значения: 'success' (успех), 'error' (ошибка),
     * 'warning' (предупреждение), 'info' (информация) или любую строку.
     */
    status?: 'success' | 'error' | 'warning' | 'info' | string

    /**
     * Заголовок результата, отображаемый крупным шрифтом.
     * Это обязательный пропс.
     */
    title: string

    /**
     * Дополнительный текст, который уточняет информацию заголовка.
     * Это необязательный пропс.
     */
    subTitle?: string

    /**
     * Дополнительные элементы, такие как кнопки, которые будут отображены
     * под заголовком и подписью.
     * Это необязательный пропс.
     */
    extra?: React.ReactNode

    /**
     * URL изображения, которое будет отображаться над заголовком.
     * Это необязательный пропс.
     */
    image?: string

    /**
     * SVG или другой элемент, который будет отображаться в качестве иконки
     * перед заголовком. Это необязательный пропс.
     */
    icon?: React.ReactNode // Изменил тип на React.ReactNode для поддержки JSX элементов
}